+++ src/compiler.js
